+++ src/main/java/com/example/kinoxpbackend/repository/CinemaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CinemaRepository extends JpaRepository<Cinema,Long> {
+public interface CinemaRepository extends JpaRepository<Cinema,Integer> {
 //    List<Cinema> findAll(Long id);
 }
